refactor(sync): extract platform_routing from orchestrator
The deprecated v0.7.0 routing surface (`SyncDirection` enum,
`parse_direction`, plus the still-supported `default_target_for`)
was buried inside the 1,159-line `orchestrator.rs`. Move it to a
dedicated `platform_routing.rs` module so:

- The deprecated surface is one grep target (`#[deprecated]` items)
  and can be deleted in one move once the migration completes.
- `orchestrator.rs` shrinks ~75 lines and reads as a single concern
  (the SyncOrchestrator type and its sync_between entrypoint).
- The 5 routing tests move with their subjects.

Public re-exports through `lib.rs` are preserved verbatim — callers
(server::handler in 4 places) build identically.

316 sync-lib tests + 2,025 workspace lib tests pass.
