+++ crates/sync/src/orchestrator.rs
@@ -8,26 +8,6 @@ use anyhow::bail;
 use serde::{Deserialize, Serialize};
 use std::collections::{HashMap, HashSet};
 
-/// Source platform for sync operation.
-///
-/// **Deprecated**: Use [`sync_between`] with string platform names instead.
-/// This enum grows quadratically with the number of platforms and will be
-/// removed in a future release.
-#[deprecated(
-    since = "0.7.0",
-    note = "Use sync_between(from, to, params) with string platform names instead"
-)]
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum SyncDirection {
-    ClaudeToCodex,
-    CodexToClaude,
-    ClaudeToCopilot,
-    CopilotToClaude,
-    ClaudeToCursor,
-    CursorToClaude,
-}
-
 /// Parameters for a sync operation.
 ///
 /// ```
@@ -497,52 +477,6 @@ impl<S: AgentAdapter, T: AgentAdapter> SyncOrchestrator<S, T> {
     }
 }
 
-/// Determines sync direction from string input (legacy API).
-///
-/// **Deprecated**: Use [`sync_between`] with [`default_target_for`] instead.
-///
-/// ```
-/// #[allow(deprecated)]
-/// use skrills_sync::{parse_direction, SyncDirection};
-///
-/// #[allow(deprecated)]
-/// {
-///     assert_eq!(parse_direction("claude").unwrap(), SyncDirection::ClaudeToCodex);
-///     assert_eq!(parse_direction("codex").unwrap(), SyncDirection::CodexToClaude);
-///     assert!(parse_direction("invalid").is_err());
-/// }
-/// ```
-#[deprecated(
-    since = "0.7.0",
-    note = "Use sync_between(from, to, params) with default_target_for(from) instead"
-)]
-#[allow(deprecated)]
-pub fn parse_direction(from: &str) -> Result<SyncDirection> {
-    match from.to_lowercase().as_str() {
-        "claude" => Ok(SyncDirection::ClaudeToCodex),
-        "codex" => Ok(SyncDirection::CodexToClaude),
-        "copilot" => Ok(SyncDirection::CopilotToClaude),
-        "cursor" => Ok(SyncDirection::CursorToClaude),
-        _ => bail!(
-            "Unknown source '{}'. Use 'claude', 'codex', 'copilot', or 'cursor'",
-            from
-        ),
-    }
-}
-
-/// Returns the default target platform for a given source.
-///
-/// Used when a sync tool needs to infer the target from only the source name.
-pub fn default_target_for(from: &str) -> &'static str {
-    match from {
-        "claude" => "codex",
-        "codex" => "claude",
-        "copilot" => "claude",
-        "cursor" => "claude",
-        _ => "codex",
-    }
-}
-
 /// Runs a sync between two named platforms using `create_adapter`.
 ///
 /// This avoids the combinatorial match arm explosion that occurs when each
@@ -761,27 +695,6 @@ mod tests {
         // Verify config was created
         let tgt_config = tgt_dir.path().join("config.json");
         assert!(tgt_config.exists());
-    }
-
-    #[test]
-    #[allow(deprecated)]
-    fn parse_direction_claude() {
-        let dir = parse_direction("claude").unwrap();
-        assert_eq!(dir, SyncDirection::ClaudeToCodex);
-    }
-
-    #[test]
-    #[allow(deprecated)]
-    fn parse_direction_codex() {
-        let dir = parse_direction("codex").unwrap();
-        assert_eq!(dir, SyncDirection::CodexToClaude);
-    }
-
-    #[test]
-    #[allow(deprecated)]
-    fn parse_direction_invalid() {
-        let result = parse_direction("invalid");
-        assert!(result.is_err());
     }
 
     #[test]
@@ -1128,22 +1041,6 @@ mod tests {
         let report = orchestrator.sync(&params).unwrap();
         assert_eq!(report.commands.written, 0);
         assert_eq!(report.commands.skipped.len(), 1);
-    }
-
-    #[test]
-    fn default_target_for_all_platforms() {
-        assert_eq!(default_target_for("claude"), "codex");
-        assert_eq!(default_target_for("codex"), "claude");
-        assert_eq!(default_target_for("copilot"), "claude");
-        assert_eq!(default_target_for("cursor"), "claude");
-        assert_eq!(default_target_for("unknown"), "codex");
-    }
-
-    #[test]
-    #[allow(deprecated)]
-    fn parse_direction_cursor() {
-        let dir = parse_direction("cursor").unwrap();
-        assert_eq!(dir, SyncDirection::CursorToClaude);
     }
 
     #[test]
